move rgb components into folder
